Add First Todo Component Form & Todo Store
Add the first Todo component form and extend the Todo store so that all
of the Todo components can use the Todo store to see the Todo list's
current state. Since there is no back to the results' storage, the Todo
store would act as the backend of sorts for this project.